fix: component 분리 및 css 수정
