+++ src/styles/ViewDetailSsul/AlreadyComplainModalComponentStyle.ts
