+++ src/styles/ViewDetailSsul/ViewDetailSsulPageComponentStyle.ts
@@ -1,16 +1,22 @@
 import styled from 'styled-components';
 
-export const DSDiv = styled.div`
+export const DSDiv = styled.div<{
+  $isRemoveModal: boolean;
+  $isComplainModalOpen: boolean;
+}>`
   display: flex;
   flex-direction: column;
   width: 100vw;
   height: 100vh;
+  overflow: ${({ $isRemoveModal, $isComplainModalOpen }) =>
+    $isRemoveModal || $isComplainModalOpen ? 'hidden' : 'overlay'};
+  box-sizing: border-box;
 `;
 
 export const DSCDiv = styled.div`
   display: flex;
   flex-direction: column;
-  width: 68%;
+  width: 80%;
   height: 100%;
   margin: 0 auto;
